put some logics into connector
