+++ source/jormungandr/jormungandr/street_network/street_network.py
@@ -45,7 +45,7 @@ class AbstractStreetNetworkService(ABC):
         pass
 
     @abc.abstractmethod
-    def direct_path(self, mode, pt_object_origin, pt_object_destination, fallback_extremity, request):
+    def direct_path(self, mode, pt_object_origin, pt_object_destination, fallback_extremity, request, **kwargs):
         '''
         :param fallback_extremity: is a PeriodExtremity (a datetime and it's meaning on the fallback period)
         '''
