+++ source/jormungandr/jormungandr/utils.py
@@ -388,3 +388,11 @@ def decode_polyline(encoded):
 #     represents_start: is True if it's start of period, False if it's the end of period
 # (mostly used for fallback management in experimental scenario)
 PeriodExtremity = namedtuple('PeriodExtremity', ['datetime', 'represents_start'])
+
+
+class SectionSorter(object):
+    def __call__(self, a, b):
+        if a.begin_date_time != b.begin_date_time:
+            return -1 if a.begin_date_time < b.begin_date_time else 1
+        else:
+            return -1 if a.end_date_time < b.end_date_time else 1
